task: combined linting action

+++ rayz/util.py
@@ -3,6 +3,6 @@ class ColumnIndexTooLarge(Exception):
 class RowIndexTooLarge(Exception):
     pass
 
-def equal(a: float, b: float) -> float:
+def equal(a: float, b: float) -> int:
     EPSILON = 0.00001
     return abs(a - b) < EPSILON
